♻️ Extract GameState
