+++ composeApp/src/commonMain/kotlin/live/ditto/demo/kmp/Ditto.kt
@@ -1,7 +1,6 @@
 package live.ditto.demo.kmp
 
 import kotlinx.coroutines.flow.Flow
-import live.ditto.demo.kmp.GameViewModel.GameState
 
 const val COLLECTION_NAME = "game_state"
 const val DOCUMENT_ID = "32760333-D9FE-4023-89C5-6646C5E3C615"
