+++ composeApp/src/commonMain/kotlin/live/ditto/demo/kmp/GameViewModel.kt
@@ -26,64 +26,6 @@ class GameViewModel(
         GRAY(Color.Gray),
     }
 
-    private val startState = listOf(
-        WHITE,
-        WHITE,
-        WHITE,
-        WHITE,
-        WHITE,
-        WHITE,
-        WHITE,
-        WHITE,
-        WHITE,
-    )
-
-    data class GameState(
-        var currentState: MutableList<GameColor> = mutableListOf(
-            WHITE,
-            WHITE,
-            WHITE,
-            WHITE,
-            WHITE,
-            WHITE,
-            WHITE,
-            WHITE,
-            WHITE,
-        )
-    ) {
-        companion object {
-            /** Deserialize GameState from a Map. */
-            fun fromMap(map: Map<String, Any?>): GameState {
-                val list = mutableListOf<GameColor>()
-                for (i in 0..8) {
-                    val stringColor = map["$i"] as String
-                    list.add(GameColor.valueOf(stringColor))
-                }
-                return GameState(list)
-            }
-        }
-
-        fun toMap(): Map<String, Any?> {
-            val map = mutableMapOf<String, Any?>()
-            currentState.toTypedArray().forEachIndexed { index, color ->
-                map["$index"] = color.name
-            }
-            return map
-        }
-
-        fun buttonTapped(buttonNumber: Int, newColor: GameColor) {
-            currentState[buttonNumber] = newColor
-            observer(this)
-        }
-
-        /** Only one observer at a time. */
-        private var observer: (GameState) -> Unit = {}
-
-        fun observe(callback: (GameState) -> Unit) {
-            observer = callback
-        }
-    }
-
     fun buttonTapped(buttonNumber: Int) {
         when (buttonNumber) {
             0 -> _button1Color.value = myColor
